ajustes en muchos componentes

+++ src/main/java/com/devs/demoCours/api/models/responses/response/CursoResponse.java
@@ -30,7 +30,6 @@ public class CursoResponse {
     @Schema(example = "PRESENCIAL")
     private ModalidadCurso modalidad;
     @Schema(example = "8")
-    private Integer duration;
+    private long duration;
     private String image;
-    //private List<SessionResponse> sessionEntityList;
 }
